// package/storefront-spiceleaf/src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
	ID: { input: string; output: string };
	String: { input: string; output: string };
	Boolean: { input: boolean; output: boolean };
	Int: { input: number; output: number };
	Float: { input: number; output: number };
	/** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
	DateTime: { input: any; output: any };
	/** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
	JSON: { input: any; output: any };
	/** The `Money` scalar type represents monetary values and supports signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
	Money: { input: any; output: any };
	/** The `Upload` scalar type represents a file upload. */
	Upload: { input: any; output: any };
};
// ...
export type Collection = Node & {
	__typename?: 'Collection';
	assets: Array<Asset>;
	breadcrumbs: Array<CollectionBreadcrumb>;
	children?: Maybe<Array<Collection>>;
	createdAt: Scalars['DateTime']['output'];
	customFields?: Maybe<Scalars['JSON']['output']>;
	description: Scalars['String']['output'];
	featuredAsset?: Maybe<Asset>;
	filters: Array<ConfigurableOperation>;
	id: Scalars['ID']['output'];
	languageCode?: Maybe<LanguageCode>;
	name: Scalars['String']['output'];
	parent?: Maybe<Collection>;
	parentId: Scalars['ID']['output'];
	position: Scalars['Int']['output'];
	productVariants: ProductVariantList;
	slug: Scalars['String']['output'];
	translations: Array<CollectionTranslation>;
	updatedAt: Scalars['DateTime']['output'];
};
// ...
export type CollectionResult = {
	__typename?: 'CollectionResult';
	collection: Collection;
	count: Scalars['Int']['output'];
};
// ...
export type CurrencyCode = (typeof CurrencyCode)[keyof typeof CurrencyCode];
// ...
export type FacetValue = Node & {
	__typename?: 'FacetValue';
	code: Scalars['String']['output'];
	createdAt: Scalars['DateTime']['output'];
	customFields?: Maybe<Scalars['JSON']['output']>;
	facet: Facet;
	facetId: Scalars['ID']['output'];
	id: Scalars['ID']['output'];
	languageCode: LanguageCode;
	name: Scalars['String']['output'];
	translations: Array<FacetValueTranslation>;
	updatedAt: Scalars['DateTime']['output'];
};
// ...
export type FacetValueResult = {
	__typename?: 'FacetValueResult';
	count: Scalars['Int']['output'];
	facetValue: FacetValue;
};
// ...
export type SearchResponse = {
	__typename?: 'SearchResponse';
	collections: Array<CollectionResult>;
	facetValues: Array<FacetValueResult>;
	items: Array<SearchResult>;
	totalItems: Scalars['Int']['output'];
};

export type Variant = {
	id: string;
	name: string;
	sku: string;
	price: number;
	priceWithTax: number;
	options: {
		name: string;
		value: string;
	}[];
};

export type SearchResult = {
	__typename?: 'SearchResult';
	/** An array of ids of the Collections in which this result appears */
	collectionIds: Array<Scalars['ID']['output']>;
	currencyCode: CurrencyCode;
	description: Scalars['String']['output'];
	facetIds: Array<Scalars['ID']['output']>;
	facetValueIds: Array<Scalars['ID']['output']>;
	inStock: Scalars['Boolean']['output'];
	price: SearchResultPrice;
	priceWithTax: SearchResultPrice;
	productAsset?: Maybe<SearchResultAsset>;
	productId: Scalars['ID']['output'];
	productName: Scalars['String']['output'];
	productVariantAsset?: Maybe<SearchResultAsset>;
	productVariantId: Scalars['ID']['output'];
	productVariantName: Scalars['String']['output'];
	/** A relevance score for the result. Differs between database implementations */
	score: Scalars['Float']['output'];
	sku: Scalars['String']['output'];
	slug: Scalars['String']['output'];
	varients: Variant[];
};

export type SearchResultAsset = {
	__typename?: 'SearchResultAsset';
	focalPoint?: Maybe<Coordinate>;
	id: Scalars['ID']['output'];
	preview: Scalars['String']['output'];
};

/** The price of a search result product, either as a range or as a single price */
export type SearchResultPrice = PriceRange | SinglePrice;